mock classes extracted
